The big refactor
Switched to gradle build system

Added version checker

Refactored classes into packages

Added remove item ui

Buy quantity options in full, half, quarter stacks and single items

New filter for oredict entries

Some fixes with entity names
