--- compile.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<project default="create_jar" name="Villager Shops (Sponge Plugin)">
-    <!--this file was created by Eclipse Runnable JAR Export Wizard-->
-    <!--ANT 1.7 is required                                        -->
-    <!--define folder properties-->
-	<property name="dir.buildfile" value="${basedir}" />
-    <property name="dir.workspace" value="${dir.buildfile}/.."/>
-    <property name="dir.server" value="G:/Sonstige/VanillaSponge/server"/>
-    <property name="dir.gpg" value="G:/Sonstige/VanillaSponge/GnuPG"/>
-    
-    <property name="my.project.name" value="set by mcmod.info" />
-    <property name="my.project.version" value="set by mcmod.info" />
-
-    <script language="javascript">
-        var File = Java.type('java.io.File');
-        var FileReader = Java.type('java.io.FileReader');
-        var BufferedReader = Java.type('java.io.BufferedReader');
-
-        var mcmod; {
-            //open mcmod.info from resources
-            var file = new File("res/mcmod.info");
-            fr = new FileReader(file);
-            br = new BufferedReader(fr);
-
-            //read whole file
-            var data = ""; var line; while (line = br.readLine()) data += line;
-
-            //JSON parse
-            mcmod=JSON.parse(data);
-        } // contained in it's own code-block for cleaner variable context
-
-        project.setProperty("my.project.version", mcmod[0].version);
-        project.setProperty("my.project.name", mcmod[0].name);
-    </script>
-
-    <target name="create_jar">
-        <jar destfile="${dir.server}/mods/${my.project.name}-${my.project.version}.jar" filesetmanifest="mergewithoutmain">
-            <fileset dir="${dir.buildfile}/bin"/>
-        </jar>
-    </target>
-    <target name="run_server">
-    	<exec dir="${dir.server}" executable="cmd">
-    		<arg line="/c start cmd /c startServer.bat"/>
-    	</exec>
-    </target>
-    <target name="build_and_run" depends="create_jar">
-    	<antcall target="run_server"/>
-    </target>
-    <target name="release_build" depends="create_jar">
-    	<exec dir="${dir.gpg}" executable="cmd">
-    		<arg line="/c start cmd /c gpg2 --output ${dir.buildfile}/${my.project.name}-${my.project.version}.jar.sig --detach-sig ${dir.server}/mods/${my.project.name}-${my.project.version}.jar"/>
-    	</exec>
-    </target>
-</project>
