+++ src/main/java/de/dosmike/sponge/vshop/DependingSuggestionElement.java
