+++ src/main/resources/assets/vshop/default_settings.conf
@@ -1,0 +1,13 @@
+# It's strongly recommended to enable automatic version checking,
+# This will also inform you about changes in dependencies.
+# Set this value to true to allow this Plugin to check for Updates on Ore
+VersionChecking = false,
+
+# In case you don't want to bother downloading the translations
+# yourself or didn't read the ore page for this plugin you can enable
+# this option to automatically download the translations from github
+AutoDownload = false,
+
+# This is the default selected stack-size for shops.
+# Allowed values are 1, 16, 32 and 64
+DefaultStackSize = 64
